moving Environment for separate module -- this will be used by different algorithms
